chore: Payment의 amount Long -> Integer 통일, import 충돌 수정

+++ backend/src/main/java/com/pickgo/domain/payment/dto/PaymentCreateRequest.java
@@ -6,14 +6,14 @@ import com.pickgo.domain.reservation.entity.Reservation;
 import jakarta.validation.constraints.NotNull;
 
 public record PaymentCreateRequest(
-		@NotNull Long amount,
-		@NotNull Long reservationId
+        @NotNull Integer amount,
+        @NotNull Long reservationId
 ) {
-	public Payment toEntity(Reservation reservation) {
-		return Payment.builder()
-				.amount(amount)
-				.status(PaymentStatus.PENDING) // 기본값으로 설정한다고 가정
-				.reservation(reservation)
-				.build();
-	}
+    public Payment toEntity(Reservation reservation) {
+        return Payment.builder()
+                .amount(amount)
+                .status(PaymentStatus.PENDING) // 기본값으로 설정한다고 가정
+                .reservation(reservation)
+                .build();
+    }
 }
