+++ backend/src/main/java/com/pickgo/domain/payment/dto/PaymentSimpleResponse.java
@@ -4,15 +4,15 @@ import com.pickgo.domain.payment.entity.Payment;
 import com.pickgo.domain.payment.entity.PaymentStatus;
 
 public record PaymentSimpleResponse(
-		Long id,
-		Long amount,
-		PaymentStatus paymentStatus
+        Long id,
+        Integer amount,
+        PaymentStatus paymentStatus
 ) {
-	public static PaymentSimpleResponse from(Payment payment) {
-		return new PaymentSimpleResponse(
-				payment.getId(),
-				payment.getAmount(),
-				payment.getStatus()
-		);
-	}
+    public static PaymentSimpleResponse from(Payment payment) {
+        return new PaymentSimpleResponse(
+                payment.getId(),
+                payment.getAmount(),
+                payment.getStatus()
+        );
+    }
 }
